update run files
