+++ lib/main_dev.dart
@@ -1,28 +1,13 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:pawdio/config.dart';
-import 'package:pawdio/screens/library/library.dart';
-
 import 'package:flutter_stetho/flutter_stetho.dart';
+import 'package:pawdio/config.dart';
+import 'package:pawdio/pawdio_app.dart';
 
 /// Dev mode includes Stetho for remote sqlite debugging at chrome://inspect/#devices
 void main() {
-  print('devmode_D E V   M O D E_devmode');
-  var configuredApp = Config(appName: 'Pawdio', flavor: 'dev', child: MyApp());
+  print('devmode__D E V   M O D E__devmode');
+  var configuredApp =
+      Config(appName: 'Pawdio', flavor: 'prod', child: PawdioApp());
   Stetho.initialize();
   runApp(configuredApp);
-}
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    // imageCache.clear();
-    return MaterialApp(
-        title: 'The best audio player in the world',
-        home: LibraryScreen(),
-        // home: Playscreen(),
-        theme: ThemeData.dark(),
-      );
-  }
 }
