+++ pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
   audioplayers: ^0.15.1
   file_picker:
   path_provider:
+  redux:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
